requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
pandas>=2.0
scikit-learn>=1.2
joblib>=1.2
pytest>=7.0
pytest-django>=4.5
python-dotenv>=1.0
